requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
ale-py==0.7.4
anyio==3.5.0
appdirs==1.4.4
arxiv==1.4.2
async-timeout==4.0.2
autograd==1.4
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
Babel==2.9.1
black==19.10b0
blitz-bayesian-pytorch==0.2.8
box2d-py==2.3.5
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==2.0.12
chex==0.1.6
cloudpickle==2.0.0
Cython==0.29.28
datasets==2.4.0
Deprecated==1.2.13
dill==0.3.5.1
dm-control==1.0.2
dm-env==1.5
dm-tree==0.1.7
docker-pycreds==0.4.0
dot2tex==2.11.3
einops==0.4.1
entrypoints==0.3
etils==1.0.0
fasteners==0.17.3
fastjsonschema==2.15.3
feedparser==6.0.10
filelock==3.4.2
fire==0.4.0
flax==0.6.6
fonttools==4.25.0
frozenlist==1.3.0
fsspec==2022.7.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
grpcio==1.43.0
gtimer==1.0.0b5
gym==0.21.0
huggingface-hub==0.8.1
hypnettorch
imageio==2.17.0
importlib-resources==5.6.0
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-server==2.13.0
labmaze==1.0.5
line-profiler==3.5.1
lockfile==0.12.2
LovelyPlots==0.0.26
lxml==4.8.0
lz4==4.0.0
Markdown==3.3.6
markdown-it-py==2.2.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
msgpack==1.0.3
mujoco==2.1.5
mujoco-py==1.50.1.68
multidict==6.0.2
multiprocess==0.70.13
munkres==1.1.4
mypy-extensions==0.4.3
natsort==8.1.0
nbclassic==0.3.7
nbconvert==6.5.0
notebook-shim==0.1.0
nptyping==2.2.0
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.2
pandas==1.3.5
panflute==2.1.3
pathspec==0.7.0
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyglet==1.5.23
Pygments==2.14.0
pylatexenc==2.10
PyOpenGL==3.1.6
pytinyrenderer==0.0.13
pytz==2021.3
PyYAML==6.0
ray==1.9.2
redis==4.1.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.1
rsa==4.8
scikit-learn==1.1.1
sentry-sdk==1.5.9
setproctitle==1.2.2
sgmllib3k==1.0.0
shortuuid==1.0.8
sip==4.19.13
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.2.0
soupsieve==2.3.2.post1
stable-baselines3==1.5.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorstore==0.1.33
termcolor==1.1.0
terminado==0.9.4
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchbnn==1.2
torchvision==0.12.0
tqdm==4.64.0
trimesh==3.20.1
typing_extensions==4.5.0
urllib3==1.26.9
wandb==0.12.16
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
wrapt==1.13.3
xxhash==3.0.0
yarl==1.7.2
